gui_requirements_full.txt
# Minimal requirements for VintageOptics GUI

# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
flask>=2.0.0
flask-cors>=3.0.0

# Scientific computing
scipy>=1.7.0

# Optional but recommended
PyWavelets>=1.1.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
